fix(plugin-data-viewer): CB-1498 move logic to the data viewer plugin
