+++ webapp/packages/plugin-object-viewer/src/ObjectViewerBootstrap.ts
@@ -6,23 +6,16 @@
  * you may not use this file except in compliance with the License.
  */
 
-import { NavigationTabsService } from '@cloudbeaver/core-app';
-import { ConnectionsManagerService, IConnectionExecutorData } from '@cloudbeaver/core-connections';
 import { Bootstrap, injectable } from '@cloudbeaver/core-di';
-import { ExecutorInterrupter, IExecutionContextProvider } from '@cloudbeaver/core-executor';
-import { DataViewerTabService } from '@cloudbeaver/plugin-data-viewer';
 
 import { ObjectPropertiesPageService } from './ObjectPropertiesPage/ObjectPropertiesPageService';
-import { isObjectViewerTab, ObjectViewerTabService } from './ObjectViewerTabService';
+import { ObjectViewerTabService } from './ObjectViewerTabService';
 
 @injectable()
 export class ObjectViewerBootstrap extends Bootstrap {
   constructor(
     private readonly objectViewerTabService: ObjectViewerTabService,
     private readonly objectPropertiesPageService: ObjectPropertiesPageService,
-    private readonly connectionsManagerService: ConnectionsManagerService,
-    private readonly navigationTabsService: NavigationTabsService,
-    private readonly dataViewerTabService: DataViewerTabService
   ) {
     super();
   }
@@ -30,29 +23,7 @@ export class ObjectViewerBootstrap extends Bootstrap {
   register(): void {
     this.objectViewerTabService.registerTabHandler();
     this.objectPropertiesPageService.registerDBObjectPage();
-
-    this.connectionsManagerService.onDisconnect.addHandler(this.disconnectHandler.bind(this));
   }
 
   load(): void { }
-
-  private async disconnectHandler(
-    data: IConnectionExecutorData,
-    contexts: IExecutionContextProvider<IConnectionExecutorData>
-  ) {
-    if (data.state === 'before') {
-      const tabs = Array.from(this.navigationTabsService.findTabs(
-        isObjectViewerTab(tab => data.connections.includes(tab.handlerState.connectionId ?? ''))
-      ));
-
-      for (const tab of tabs) {
-        const canClose = await this.dataViewerTabService.canClose(tab);
-
-        if (!canClose) {
-          ExecutorInterrupter.interrupt(contexts);
-          return;
-        }
-      }
-    }
-  }
 }
